Move copy into clipboard functionality into its own helper
